Change from specific avoidance status message to a more generic companion status message
