--- mavros_msgs/msg/AvoidanceStatus.msg
@@ -1,12 +1,0 @@
-# MAVLink message: AVOIDANCESTATUS
-# http://mavlink.org/messages/common#AVOIDANCESTATUS
-
-std_msgs/Header header
-
-uint8 status # See enum OBS_AVOID_STATUS
-
-uint8 OBS_AVOID_STATUS_HEALTHY = 0
-uint8 OBS_AVOID_STATUS_TIMEOUT = 1
-uint8 OBS_AVOID_STATUS_ABORT = 2
-uint8 OBS_AVOID_STATUS_NOT_READY = 3
-uint8 OBS_AVOID_STATUS_CAMERA_FAIL = 4
